Add example of test
Uses supertests and jest.

Based on https://www.albertgao.xyz/2017/05/24/how-to-test-expressjs-with-jest-and-supertest/
