+++ .travis.yml
@@ -9,3 +9,4 @@ install:
 
 script:
   - npm run lint
+  - npm run test
